requirements.txt
streamlit==1.41.0
plotly==5.24.0
pandas==2.2.3
gspread==6.1.4
google-auth==2.37.0
google-analytics-data==0.18.15
requests==2.32.3
python-dotenv==1.0.1
schedule==1.2.2
